Remove DruidWindowQueryTest and WindowQueryTestBase, move those tests to DrillWindowQueryTest
