+++ sql/src/test/resources/drill/window/queries/druid_queries/multiple_windows/wikipedia_query_1.e
