+++ sql/src/test/resources/drill/window/queries/druid_queries/shuffle_columns/wikipedia_query_2.e
